CardTokenization* migrated to CardPayment* in example app
