--- example/src/main/kotlin/com/processout/example/ui/screen/card/InvoiceDetails.kt
@@ -1,6 +1,0 @@
-package com.processout.example.ui.screen.card
-
-data class InvoiceDetails(
-    val amount: String,
-    val currency: String
-)
