+++ example/src/main/kotlin/com/processout/example/ui/screen/card/payment/CardPaymentUiState.kt
@@ -1,12 +1,13 @@
 package com.processout.example.ui.screen.card.payment
 
-import com.processout.example.ui.screen.card.InvoiceDetails
 import com.processout.sdk.core.ProcessOutResult
 
 sealed class CardPaymentUiState {
     data object Initial : CardPaymentUiState()
     data object Submitting : CardPaymentUiState()
     data class Submitted(val uiModel: CardPaymentUiModel) : CardPaymentUiState()
+    data class Tokenizing(val uiModel: CardPaymentUiModel) : CardPaymentUiState()
+    data class Tokenized(val uiModel: CardPaymentUiModel) : CardPaymentUiState()
     data class Authorizing(val uiModel: CardPaymentUiModel) : CardPaymentUiState()
     data class Failure(val failure: ProcessOutResult.Failure) : CardPaymentUiState()
 }
@@ -14,17 +15,10 @@ sealed class CardPaymentUiState {
 data class CardPaymentUiModel(
     val cardId: String,
     val invoiceId: String,
-    val details: CardPaymentDetails
-)
-
-data class CardPaymentDetails(
-    val card: CardDetails,
     val invoice: InvoiceDetails
 )
 
-data class CardDetails(
-    val number: String,
-    val expMonth: String,
-    val expYear: String,
-    val cvc: String
+data class InvoiceDetails(
+    val amount: String,
+    val currency: String
 )
