+++ example/src/main/kotlin/com/processout/example/ui/screen/card/payment/CardPaymentViewModel.kt
@@ -4,75 +4,50 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
 import com.processout.example.shared.Constants
-import com.processout.example.ui.screen.card.InvoiceDetails
 import com.processout.example.ui.screen.card.payment.CardPaymentUiState.*
 import com.processout.sdk.api.ProcessOut
-import com.processout.sdk.api.model.request.POCardTokenizationRequest
 import com.processout.sdk.api.model.request.POCreateInvoiceRequest
-import com.processout.sdk.api.repository.POCardsRepository
+import com.processout.sdk.api.model.response.POCard
 import com.processout.sdk.api.service.POInvoicesService
-import com.processout.sdk.core.getOrNull
 import com.processout.sdk.core.onFailure
-import kotlinx.coroutines.async
+import com.processout.sdk.core.onSuccess
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 import java.util.UUID
 
 class CardPaymentViewModel(
-    private val cards: POCardsRepository,
     private val invoices: POInvoicesService
 ) : ViewModel() {
 
     class Factory : ViewModelProvider.Factory {
         @Suppress("UNCHECKED_CAST")
         override fun <T : ViewModel> create(modelClass: Class<T>): T =
-            with(ProcessOut.instance) {
-                CardPaymentViewModel(
-                    cards,
-                    invoices
-                ) as T
-            }
+            CardPaymentViewModel(
+                ProcessOut.instance.invoices
+            ) as T
     }
 
     private val _uiState = MutableStateFlow<CardPaymentUiState>(Initial)
     val uiState = _uiState.asStateFlow()
 
-    fun submit(details: CardPaymentDetails) {
+    fun submit(details: InvoiceDetails) {
         _uiState.value = Submitting
         viewModelScope.launch {
-            val cardDeferred = async { tokenize(details.card) }
-            val invoiceDeferred = async { createInvoice(details.invoice) }
-            val cardResult = cardDeferred.await()
-            val invoiceResult = invoiceDeferred.await()
-
-            cardResult.onFailure { _uiState.value = Failure(it) }
-            invoiceResult.onFailure { _uiState.value = Failure(it) }
-
-            val cardId = cardResult.getOrNull()?.id
-            val invoiceId = invoiceResult.getOrNull()?.id
-            if (cardId != null && invoiceId != null) {
-                _uiState.value = Submitted(
-                    CardPaymentUiModel(
-                        cardId = cardId,
-                        invoiceId = invoiceId,
-                        details = details
-                    )
-                )
+            createInvoice(details).let { result ->
+                result.onFailure { _uiState.value = Failure(it) }
+                    .onSuccess {
+                        _uiState.value = Submitted(
+                            CardPaymentUiModel(
+                                cardId = String(),
+                                invoiceId = it.id,
+                                invoice = details
+                            )
+                        )
+                    }
             }
         }
     }
-
-    private suspend fun tokenize(details: CardDetails) =
-        cards.tokenize(
-            POCardTokenizationRequest(
-                name = "John Doe",
-                number = details.number,
-                expMonth = details.expMonth.toInt(),
-                expYear = details.expYear.toInt(),
-                cvc = details.cvc
-            )
-        )
 
     private suspend fun createInvoice(details: InvoiceDetails) =
         invoices.createInvoice(
@@ -84,10 +59,28 @@ class CardPaymentViewModel(
             )
         )
 
-    fun onAuthorizing() {
+    fun onTokenizing() {
         val uiState = _uiState.value
         if (uiState is Submitted) {
-            _uiState.value = Authorizing(uiState.uiModel)
+            _uiState.value = Tokenizing(uiState.uiModel.copy())
+        }
+    }
+
+    fun onTokenized(card: POCard) {
+        val uiState = _uiState.value
+        if (uiState is Tokenizing) {
+            _uiState.value = Tokenized(
+                uiState.uiModel.copy(
+                    cardId = card.id
+                )
+            )
+        }
+    }
+
+    fun onAuthorizing() {
+        val uiState = _uiState.value
+        if (uiState is Tokenized) {
+            _uiState.value = Authorizing(uiState.uiModel.copy())
         }
     }
