+++ example/src/main/kotlin/com/processout/example/ui/screen/features/FeaturesFragment.kt
@@ -39,11 +39,6 @@ class FeaturesFragment : BaseFragment<FragmentFeaturesBinding>(
         super.onViewCreated(view, savedInstanceState)
         val navController = findNavController()
         with(binding) {
-            cardTokenizationButton.setOnClickListener {
-                navController.navigate(
-                    FeaturesFragmentDirections.actionFeaturesFragmentToCardTokenizationFragment()
-                )
-            }
             cardPaymentButton.setOnClickListener {
                 navController.navigate(
                     FeaturesFragmentDirections.actionFeaturesFragmentToCardPaymentFragment()
