--- example/src/main/res/layout/fragment_card_tokenization.xml
@@ -1,139 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.screen.card.payment.CardPaymentFragment">
-
-    <androidx.core.widget.NestedScrollView
-        android:id="@+id/scrollable_content"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:defaultFocusHighlightEnabled="false"
-        android:fillViewport="true"
-        android:isScrollContainer="true"
-        app:layout_constraintBottom_toTopOf="@id/footer_divider"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        tools:targetApi="O">
-
-        <LinearLayout
-            android:id="@+id/inputs_container"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical"
-            android:paddingHorizontal="@dimen/layout_padding"
-            android:paddingTop="@dimen/content_padding"
-            android:paddingBottom="@dimen/content_marginVertical"
-            app:layout_constraintLeft_toLeftOf="parent"
-            app:layout_constraintRight_toRightOf="parent"
-            app:layout_constraintTop_toTopOf="parent">
-
-            <TextView
-                android:id="@+id/invoice_details_subtitle"
-                style="@style/Widget.ProcessOut.Subtitle"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="@string/invoice_details" />
-
-            <com.google.android.material.textfield.TextInputLayout
-                android:id="@+id/amount_input_layout"
-                style="@style/Widget.ProcessOut.Example.TextInputLayout"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="@dimen/subtitle_paddingVertical"
-                android:hint="@string/amount">
-
-                <com.google.android.material.textfield.TextInputEditText
-                    android:id="@+id/amount_input"
-                    style="@style/Widget.ProcessOut.Example.EditText"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:importantForAutofill="no"
-                    android:inputType="numberDecimal"
-                    android:text="5" />
-            </com.google.android.material.textfield.TextInputLayout>
-
-            <com.google.android.material.textfield.TextInputLayout
-                android:id="@+id/currency_input_layout"
-                style="@style/Widget.ProcessOut.Example.TextInputLayout"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="@dimen/input_space_vertical"
-                android:hint="@string/currency">
-
-                <com.google.android.material.textfield.TextInputEditText
-                    android:id="@+id/currency_input"
-                    style="@style/Widget.ProcessOut.Example.EditText"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:imeOptions="actionDone"
-                    android:importantForAutofill="no"
-                    android:inputType="text|textCapCharacters"
-                    android:text="USD" />
-            </com.google.android.material.textfield.TextInputLayout>
-
-            <TextView
-                android:id="@+id/threeds_service_subtitle"
-                style="@style/Widget.ProcessOut.Subtitle"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="@dimen/layout_padding"
-                android:text="@string/threeds_service" />
-
-            <RadioGroup
-                android:id="@+id/threeds_service_radio_group"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:checkedButton="@id/threeds_service_test">
-
-                <RadioButton
-                    android:id="@+id/threeds_service_test"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:text="@string/threeds_service_test"
-                    android:textAppearance="@style/TextAppearance.ProcessOut.Fixed.Body" />
-
-                <RadioButton
-                    android:id="@+id/threeds_service_checkout"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:text="@string/threeds_service_checkout"
-                    android:textAppearance="@style/TextAppearance.ProcessOut.Fixed.Body" />
-
-                <RadioButton
-                    android:id="@+id/threeds_service_adyen"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:text="@string/threeds_service_adyen"
-                    android:textAppearance="@style/TextAppearance.ProcessOut.Fixed.Body" />
-            </RadioGroup>
-        </LinearLayout>
-    </androidx.core.widget.NestedScrollView>
-
-    <View
-        android:id="@+id/footer_divider"
-        android:layout_width="match_parent"
-        android:layout_height="@dimen/po_borderWidth"
-        android:layout_marginBottom="@dimen/button_marginVertical"
-        android:background="@color/po_border_subtle"
-        app:layout_constraintBottom_toTopOf="@+id/authorize_invoice_button"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent" />
-
-    <Button
-        android:id="@+id/authorize_invoice_button"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/content_padding"
-        android:layout_marginEnd="@dimen/content_padding"
-        android:layout_marginBottom="@dimen/button_marginVertical"
-        android:text="@string/authorize_invoice"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
