+++ example/src/main/res/values/strings.xml
@@ -11,7 +11,6 @@
 
     <string name="card_payment" translatable="false">Card Payment</string>
     <string name="card_update" translatable="false">Card Update</string>
-    <string name="card_tokenization" translatable="false">Card Tokenization (WIP)</string>
     <string name="card_details" translatable="false">Card Details</string>
     <string name="card_number" translatable="false">Number</string>
     <string name="card_expiration_month" translatable="false">Expiration Month</string>
